not needed gui classes removed

+++ dispatch-rider/src/main/java/gui/holon/HolonTableModel.java
@@ -10,9 +10,7 @@ import java.util.HashMap;
 
 public class HolonTableModel extends AbstractTableModel implements Updateable {
 
-
     HolonUpdateable updateable = new HolonUpdateable();
-
 
     @SuppressWarnings("unchecked")
     @Override
@@ -35,7 +33,7 @@ public class HolonTableModel extends AbstractTableModel implements Updateable {
         if (updateable.visualisedRecord.getData() == null) return 0;
         @SuppressWarnings("unchecked")
         HashMap<Integer, HashMap<String, Object>> extracted = (HashMap<Integer, HashMap<String, Object>>) updateable.visualisedRecord.getData();
-        return extracted.get((Integer) updateable.getHolonIds().toArray()[rowIndex]).get(updateable.getColumnNames()[columnIndex]);
+        return extracted.get(updateable.getHolonIds().toArray()[rowIndex]).get(updateable.getColumnNames()[columnIndex]);
     }
 
     public void setSimInfo(SimInfo simInfo) {
@@ -45,8 +43,6 @@ public class HolonTableModel extends AbstractTableModel implements Updateable {
     /**
      * Do aktualizowania danych w tabelach
      *
-     * @param data
-     * @param simInfo
      */
     public void update(SimulationData data) {
         updateable.update(data);
@@ -55,7 +51,6 @@ public class HolonTableModel extends AbstractTableModel implements Updateable {
     /**
      * Wskazuje ze update'y przychodza juz na nowy timestamp
      *
-     * @param val
      */
     public void newTimestampUpdate(int val) {
         updateable.newTimestampUpdate(val);
@@ -67,8 +62,6 @@ public class HolonTableModel extends AbstractTableModel implements Updateable {
 
     /**
      * Wyswietla dane z wybranego timestamp
-     *
-     * @param val
      */
     public void setDrawnTimestamp(int val) {
         updateable.setDrawnTimestamp(val);
